requirements.txt
openai
python-dotenv
colorama
google-genai
